+++ alacritty/src/window.rs
@@ -35,6 +35,12 @@ use glutin::dpi::{PhysicalPosition, PhysicalSize};
 use glutin::event_loop::EventLoop;
 #[cfg(target_os = "macos")]
 use glutin::platform::macos::{WindowBuilderExtMacOS, WindowExtMacOS};
+#[cfg(target_os = "macos")]
+use raw_window_handle::{HasRawWindowHandle, RawWindowHandle};
+#[cfg(target_os = "macos")]
+use objc::{msg_send, sel, sel_impl};
+#[cfg(target_os = "macos")]
+use cocoa::base::id;
 #[cfg(windows)]
 use glutin::platform::windows::IconExtWindows;
 use glutin::window::{
@@ -433,6 +439,19 @@ impl Window {
 
     pub fn resize(&self, size: PhysicalSize<u32>) {
         self.windowed_context.resize(size);
+    }
+
+    /// Force macOS to clear shadow of transparent windows.
+    #[cfg(target_os = "macos")]
+    pub fn invalidate_shadow(&self) {
+        let raw_window = match self.window().raw_window_handle() {
+            RawWindowHandle::MacOS(handle) => handle.ns_window as id,
+            _ => return,
+        };
+
+        unsafe {
+            let _: () = msg_send![raw_window, invalidateShadow];
+        }
     }
 
     fn window(&self) -> &GlutinWindow {